tests/requirements.txt
python-rtmidi
pyusb
pytest